+++ lib/globals.dart
@@ -20,8 +20,8 @@ final GlobalKey<ScaffoldMessengerState> scaffoldMessengerKey =
 final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
 const String prgName = 'easysend';
-const String progVersion = '0.4.260814';
-const int buildNumber = 108;
+const String progVersion = '0.4.260816';
+const int buildNumber = 109;
 const String progAuthor = 'Eugen';
 
 const String langFile = 'assets/locales.json';